unread [nfc]: Move the "direct" selectors into model file.
This is another step toward encapsulating the details of this
part of our state from the rest of the app.
